# week7/backendw7/openapi_server/models/error_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class ErrorResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, message=None, code=None, details=None):  # noqa: E501
        """ErrorResponse - a model defined in OpenAPI

        :param message: The message of this ErrorResponse.  # noqa: E501
        :type message: str
        :param code: The code of this ErrorResponse.  # noqa: E501
        :type code: str
        :param details: The details of this ErrorResponse.  # noqa: E501
        :type details: Dict[str, object]
        """
        self.openapi_types = {
            'message': str,
            'code': str,
            'details': Dict[str, object]
        }

        self.attribute_map = {
            'message': 'message',
            'code': 'code',
            'details': 'details'
        }

        self._message = message
        self._code = code
        self._details = details

    @classmethod
    def from_dict(cls, dikt) -> 'ErrorResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ErrorResponse of this ErrorResponse.  # noqa: E501
        :rtype: ErrorResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def message(self) -> str:
        """Gets the message of this ErrorResponse.


        :return: The message of this ErrorResponse.
        :rtype: str
        """
        return self._message

    @message.setter
    def message(self, message: str):
        """Sets the message of this ErrorResponse.


        :param message: The message of this ErrorResponse.
        :type message: str
        """
        if message is None:
            raise ValueError("Invalid value for `message`, must not be `None`")  # noqa: E501

        self._message = message

    @property
    def code(self) -> str:
        """Gets the code of this ErrorResponse.


        :return: The code of this ErrorResponse.
        :rtype: str
        """
        return self._code

    @code.setter
    def code(self, code: str):
        """Sets the code of this ErrorResponse.


        :param code: The code of this ErrorResponse.
        :type code: str
        """

        self._code = code

    @property
    def details(self) -> Dict[str, object]:
        """Gets the details of this ErrorResponse.


        :return: The details of this ErrorResponse.
        :rtype: Dict[str, object]
        """
        return self._details

    @details.setter
    def details(self, details: Dict[str, object]):
        """Sets the details of this ErrorResponse.


        :param details: The details of this ErrorResponse.
        :type details: Dict[str, object]
        """

        self._details = details