# week7/backendw7/openapi_server/models/product_create_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class ProductCreateRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, description=None, price=None, stock=None, category=None, is_active=True, tags=None):  # noqa: E501
        """ProductCreateRequest - a model defined in OpenAPI

        :param name: The name of this ProductCreateRequest.  # noqa: E501
        :type name: str
        :param description: The description of this ProductCreateRequest.  # noqa: E501
        :type description: str
        :param price: The price of this ProductCreateRequest.  # noqa: E501
        :type price: float
        :param stock: The stock of this ProductCreateRequest.  # noqa: E501
        :type stock: int
        :param category: The category of this ProductCreateRequest.  # noqa: E501
        :type category: str
        :param is_active: The is_active of this ProductCreateRequest.  # noqa: E501
        :type is_active: bool
        :param tags: The tags of this ProductCreateRequest.  # noqa: E501
        :type tags: List[str]
        """
        self.openapi_types = {
            'name': str,
            'description': str,
            'price': float,
            'stock': int,
            'category': str,
            'is_active': bool,
            'tags': List[str]
        }

        self.attribute_map = {
            'name': 'name',
            'description': 'description',
            'price': 'price',
            'stock': 'stock',
            'category': 'category',
            'is_active': 'isActive',
            'tags': 'tags'
        }

        self._name = name
        self._description = description
        self._price = price
        self._stock = stock
        self._category = category
        self._is_active = is_active
        self._tags = tags

    @classmethod
    def from_dict(cls, dikt) -> 'ProductCreateRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProductCreateRequest of this ProductCreateRequest.  # noqa: E501
        :rtype: ProductCreateRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this ProductCreateRequest.


        :return: The name of this ProductCreateRequest.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this ProductCreateRequest.


        :param name: The name of this ProductCreateRequest.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501
        if name is not None and len(name) > 120:
            raise ValueError("Invalid value for `name`, length must be less than or equal to `120`")  # noqa: E501
        if name is not None and len(name) < 1:
            raise ValueError("Invalid value for `name`, length must be greater than or equal to `1`")  # noqa: E501

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this ProductCreateRequest.


        :return: The description of this ProductCreateRequest.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this ProductCreateRequest.


        :param description: The description of this ProductCreateRequest.
        :type description: str
        """
        if description is not None and len(description) > 2000:
            raise ValueError("Invalid value for `description`, length must be less than or equal to `2000`")  # noqa: E501

        self._description = description

    @property
    def price(self) -> float:
        """Gets the price of this ProductCreateRequest.


        :return: The price of this ProductCreateRequest.
        :rtype: float
        """
        return self._price

    @price.setter
    def price(self, price: float):
        """Sets the price of this ProductCreateRequest.


        :param price: The price of this ProductCreateRequest.
        :type price: float
        """
        if price is None:
            raise ValueError("Invalid value for `price`, must not be `None`")  # noqa: E501
        if price is not None and price < 0:  # noqa: E501
            raise ValueError("Invalid value for `price`, must be a value greater than or equal to `0`")  # noqa: E501

        self._price = price

    @property
    def stock(self) -> int:
        """Gets the stock of this ProductCreateRequest.


        :return: The stock of this ProductCreateRequest.
        :rtype: int
        """
        return self._stock

    @stock.setter
    def stock(self, stock: int):
        """Sets the stock of this ProductCreateRequest.


        :param stock: The stock of this ProductCreateRequest.
        :type stock: int
        """
        if stock is None:
            raise ValueError("Invalid value for `stock`, must not be `None`")  # noqa: E501
        if stock is not None and stock < 0:  # noqa: E501
            raise ValueError("Invalid value for `stock`, must be a value greater than or equal to `0`")  # noqa: E501

        self._stock = stock

    @property
    def category(self) -> str:
        """Gets the category of this ProductCreateRequest.


        :return: The category of this ProductCreateRequest.
        :rtype: str
        """
        return self._category

    @category.setter
    def category(self, category: str):
        """Sets the category of this ProductCreateRequest.


        :param category: The category of this ProductCreateRequest.
        :type category: str
        """
        if category is None:
            raise ValueError("Invalid value for `category`, must not be `None`")  # noqa: E501
        if category is not None and len(category) > 80:
            raise ValueError("Invalid value for `category`, length must be less than or equal to `80`")  # noqa: E501
        if category is not None and len(category) < 1:
            raise ValueError("Invalid value for `category`, length must be greater than or equal to `1`")  # noqa: E501

        self._category = category

    @property
    def is_active(self) -> bool:
        """Gets the is_active of this ProductCreateRequest.


        :return: The is_active of this ProductCreateRequest.
        :rtype: bool
        """
        return self._is_active

    @is_active.setter
    def is_active(self, is_active: bool):
        """Sets the is_active of this ProductCreateRequest.


        :param is_active: The is_active of this ProductCreateRequest.
        :type is_active: bool
        """

        self._is_active = is_active

    @property
    def tags(self) -> List[str]:
        """Gets the tags of this ProductCreateRequest.


        :return: The tags of this ProductCreateRequest.
        :rtype: List[str]
        """
        return self._tags

    @tags.setter
    def tags(self, tags: List[str]):
        """Sets the tags of this ProductCreateRequest.


        :param tags: The tags of this ProductCreateRequest.
        :type tags: List[str]
        """
        if tags is not None and len(tags) > 30:
            raise ValueError("Invalid value for `tags`, number of items must be less than or equal to `30`")  # noqa: E501

        self._tags = tags