# week7/backendw7/openapi_server/models/product_list_response.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server.models.product import Product
from openapi_server.models.product_list_response_pagination import ProductListResponsePagination
from openapi_server import util

from openapi_server.models.product import Product  # noqa: E501
from openapi_server.models.product_list_response_pagination import ProductListResponsePagination  # noqa: E501

class ProductListResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, data=None, pagination=None):  # noqa: E501
        """ProductListResponse - a model defined in OpenAPI

        :param data: The data of this ProductListResponse.  # noqa: E501
        :type data: List[Product]
        :param pagination: The pagination of this ProductListResponse.  # noqa: E501
        :type pagination: ProductListResponsePagination
        """
        self.openapi_types = {
            'data': List[Product],
            'pagination': ProductListResponsePagination
        }

        self.attribute_map = {
            'data': 'data',
            'pagination': 'pagination'
        }

        self._data = data
        self._pagination = pagination

    @classmethod
    def from_dict(cls, dikt) -> 'ProductListResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProductListResponse of this ProductListResponse.  # noqa: E501
        :rtype: ProductListResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> List[Product]:
        """Gets the data of this ProductListResponse.


        :return: The data of this ProductListResponse.
        :rtype: List[Product]
        """
        return self._data

    @data.setter
    def data(self, data: List[Product]):
        """Sets the data of this ProductListResponse.


        :param data: The data of this ProductListResponse.
        :type data: List[Product]
        """
        if data is None:
            raise ValueError("Invalid value for `data`, must not be `None`")  # noqa: E501

        self._data = data

    @property
    def pagination(self) -> ProductListResponsePagination:
        """Gets the pagination of this ProductListResponse.


        :return: The pagination of this ProductListResponse.
        :rtype: ProductListResponsePagination
        """
        return self._pagination

    @pagination.setter
    def pagination(self, pagination: ProductListResponsePagination):
        """Sets the pagination of this ProductListResponse.


        :param pagination: The pagination of this ProductListResponse.
        :type pagination: ProductListResponsePagination
        """
        if pagination is None:
            raise ValueError("Invalid value for `pagination`, must not be `None`")  # noqa: E501

        self._pagination = pagination